Moved the proxy code into a plugin
